[admin] Add tests for multitenancy #2
Created a multitenant test for the registered models in admin.
RadiusUserGroup, RadiusGroupCheck and RadiusGroupReply don't have an
organization field through which one can check for multitenancy but
these models have as multitenant_parent group (RadiusGroup). therefore
one can test for multitenancy through the groups. In other words, an
operator belonging to org1 can only see RadiusUserGroup(s) or
RadiusGroupCheck(s) or RadiusGroupReply(ies) which belongs to RadiusGroup which
are part of org1.

Since the User and Group fields in the add_views of models which are
involved with Users and Groups are autocomplete rather than select,
testing for multitenancy involved accessing the add_view, getting the
autocomplete link for Users or Groups and performing a get on this link
to verify that operator is able to see only Users and Groups which are
members of his organization.

Fixes #2