[fix] Added validation of phone number uniqueness
If a user tries to register with username and phone number which
are already in use, only the username error is shown because
django rest auth has a validation method for this field,
but for the phone number uniqueness we rely on model validation,
which is not performed if the serializer level validation fails.

Adding a validation level serialzier for phone number too,
so that both username and phone_number are incldued in
the error response.