[radius] Add max_length validation for called_station_id and calling_station_id in PostAuthSerializer #467
Updated the PostAuthSerializer to include a `max_length` attribute of 50 for both the `called_station_id` and `calling_station_id` fields. This ensures that requests exceeding the character limit for either field return an HTTP 400 error with an appropriate error message. Added tests to validate the max_length constraint for both fields.

Fixes #467