[chores] Handled organization agnostic Accounting-On requests #617
Accounting-On/Off requests may not be tied to a specific organization.
Before we were just ignoring these requests.
Now we check the IP is coming from one of authorized IP addresses,
Accounting-Off requests are still ignored but Accounting-On
requests will close stale sessions, regardless of whether
the request is tied to a specific organization or not.

Related to #617